ETL-PROYECTO-API/simpsons_data_/requirements.txt

pandas
requests
sqlalchemy
psycopg2-binary
streamlit
apache-airflow
scikit-learn
jupyter
matplotlib
